First round of renaming ddpdemo to dfmodules
